dto type change [안세용]

+++ src/main/java/com/greedy/goodeat/common/dto/AddfileDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class AddfileDTO {
 
-	private Long addfileNo;
+	private Integer addfileNo;
 	private String originalFileName;
 	private String savedFileName;
 	private String savedRoute;
